+++ ante/fee.go
@@ -6,12 +6,13 @@ import (
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	evidencetypes "github.com/cosmos/cosmos-sdk/x/evidence/types"
 
+	appparams "github.com/umee-network/umee/v3/app/params"
 	leveragetypes "github.com/umee-network/umee/v3/x/leverage/types"
 	oracletypes "github.com/umee-network/umee/v3/x/oracle/types"
 )
 
 // MaxMsgGasUsage defines the maximum gas allowed for an oracle transaction.
-const MaxMsgGasUsage = uint64(100000)
+const MaxMsgGasUsage = uint64(100_000)
 
 // FeeAndPriority ensures tx has enough fee coins to pay for the gas at the CheckTx time
 // to early remove transactions from the mempool without enough attached fee.
@@ -24,37 +25,55 @@ func FeeAndPriority(ctx sdk.Context, tx sdk.Tx) (sdk.Coins, int64, error) {
 		return nil, 0, sdkerrors.ErrTxDecode.Wrap("Tx must be a FeeTx")
 	}
 
-	feeCoins := feeTx.GetFee()
-	gas := feeTx.GetGas()
+	providedFees := feeTx.GetFee()
+	gasLimit := feeTx.GetGas()
 	msgs := feeTx.GetMsgs()
 	isOracleOrGravity := IsOracleOrGravityTx(msgs)
-	chargeFees := !isOracleOrGravity || gas > uint64(len(msgs))*MaxMsgGasUsage
-
-	if ctx.IsCheckTx() && chargeFees {
-		minGasPrices := ctx.MinGasPrices()
-		if !minGasPrices.IsZero() {
-			requiredFees := make(sdk.Coins, len(minGasPrices))
-
-			// Determine the required fees by multiplying each required minimum gas
-			// price by the gas limit, where fee = ceil(minGasPrice * gasLimit).
-			glDec := sdk.NewDec(int64(gas))
-			for i, gp := range minGasPrices {
-				fee := gp.Amount.Mul(glDec)
-				requiredFees[i] = sdk.NewCoin(gp.Denom, fee.Ceil().RoundInt())
-			}
-
-			if !feeCoins.IsAnyGTE(requiredFees) {
-				return nil, 0, sdkerrors.ErrInsufficientFee.Wrapf(
-					"insufficient fees; got: %s required: %s", feeCoins, requiredFees)
-			}
-		}
-	}
-
 	priority := getTxPriority(isOracleOrGravity, msgs)
-	if !chargeFees {
+	chargeFees := !isOracleOrGravity || gasLimit > uint64(len(msgs))*MaxMsgGasUsage
+	// we also don't charge transaction fees for the first block, for the genesis transactions.
+	if !chargeFees || ctx.BlockHeight() == 0 {
 		return sdk.Coins{}, priority, nil
 	}
-	return feeCoins, priority, nil
+
+	var err error
+	if ctx.IsCheckTx() {
+		err = checkFees(ctx.MinGasPrices(), providedFees, gasLimit)
+	} else {
+		err = checkFees(nil, providedFees, gasLimit)
+	}
+	if err != nil {
+		err = sdkerrors.Wrap(err, msgs[0].String())
+	}
+	return providedFees, priority, err
+}
+
+func checkFees(minGasPrices sdk.DecCoins, fees sdk.Coins, gasLimit uint64) error {
+	if minGasPrices != nil {
+		// check minGasPrices set by validator
+		if err := AssertMinProtocolGasPrice(minGasPrices); err != nil {
+			return err
+		}
+	} else {
+		// in deliverTx = use protocol min gas price
+		minGasPrices = sdk.DecCoins{appparams.MinMinGasPrice}
+	}
+
+	requiredFees := make(sdk.Coins, len(minGasPrices))
+
+	// Determine the required fees by multiplying each required minimum gas
+	// price by the gas limit, where fee = ceil(minGasPrice * gasLimit).
+	glDec := sdk.NewDec(int64(gasLimit))
+	for i, gp := range minGasPrices {
+		fee := gp.Amount.Mul(glDec)
+		requiredFees[i] = sdk.NewCoin(gp.Denom, fee.Ceil().RoundInt())
+	}
+
+	if !fees.IsAnyGTE(requiredFees) {
+		return sdkerrors.ErrInsufficientFee.Wrapf(
+			"insufficient fees; got: %s required: %s", fees, requiredFees)
+	}
+	return nil
 }
 
 // IsOracleOrGravityTx checks if all messages are oracle messages
@@ -68,9 +87,8 @@ func IsOracleOrGravityTx(msgs []sdk.Msg) bool {
 			*oracletypes.MsgAggregateExchangeRateVote:
 			continue
 
-		// TODO: remove messages which should not be "free":
+		// TODO: revisit free gravity msg set
 		case *gbtypes.MsgValsetConfirm,
-			*gbtypes.MsgRequestBatch,
 			*gbtypes.MsgConfirmBatch,
 			*gbtypes.MsgERC20DeployedClaim,
 			*gbtypes.MsgConfirmLogicCall,
@@ -88,6 +106,21 @@ func IsOracleOrGravityTx(msgs []sdk.Msg) bool {
 	}
 
 	return true
+}
+
+// AssertMinProtocolGasPrice returns an error if the provided gasPrices are lower then
+// the required by protocol.
+func AssertMinProtocolGasPrice(gasPrices sdk.DecCoins) error {
+	for _, c := range gasPrices {
+		if c.Denom == appparams.MinMinGasPrice.Denom {
+			if c.Amount.LT(appparams.MinMinGasPrice.Amount) {
+				break // go to error below
+			}
+			return nil
+		}
+	}
+	return sdkerrors.ErrInsufficientFee.Wrapf(
+		"gas price too small; got: %v required min: %v", gasPrices, appparams.MinMinGasPrice)
 }
 
 // getTxPriority returns naive tx priority based on the lowest fee amount (regardless of the